testing a simpler parser that can extract signatures
I want to be able to extract a lookup of signatures for
functions, classes, and params, and we can then see if
this could be useful to assess compatibility. The next
steps for that are to add a conda parser (to get versions
from conda-forge) and/or to try doing a trace for a simple
module that uses a database we extracted to assess compatibility.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>